+++ Shared/Actions/AssetsManagement/GetResourceParams.cs
@@ -159,9 +159,15 @@ namespace CloudinaryDotNet.Actions
         public bool? Moderation { get; set; }
 
         /// <summary>
-        /// Gets or sets a value indicating whether to include accessibility analysis information. Default: false.
+        /// Gets or sets a value indicating whether to include accessibility analysis information.
+        /// Optional (Boolean, default: false).
         /// </summary>
         public bool? AccessibilityAnalysis { get; set; }
+
+        /// <summary>
+        /// Gets or sets a value indicating whether to include asset version information. Default: false.
+        /// </summary>
+        public bool? Versions { get; set; }
 
         /// <summary>
         /// Validate object model.
@@ -205,6 +211,7 @@ namespace CloudinaryDotNet.Actions
             AddParam(dict, "start_at", StartAt);
             AddParam(dict, "direction", Direction);
             AddParam(dict, "accessibility_analysis", AccessibilityAnalysis);
+            AddParam(dict, "versions", Versions);
 
             return dict;
         }