+++ Shared/Actions/PresetsUpload/UploadPresetParams.cs
@@ -376,6 +376,17 @@
         public bool Live { get; set; }
 
         /// <summary>
+        /// Gets or sets JavaScript code expression to be evaluated.
+        /// </summary>
+        public string Eval { get; set; }
+
+        /// <summary>
+        /// Gets or sets accessibility analysis information.
+        /// Optional (Boolean, default: false).
+        /// </summary>
+        public bool? AccessibilityAnalysis { get; set; }
+
+        /// <summary>
         /// Validate object model.
         /// </summary>
         public override void Check()
@@ -425,6 +436,8 @@
             AddParam(dict, "overwrite", Overwrite);
             AddParam(dict, "unique_filename", UniqueFilename);
             AddParam(dict, "live", Live);
+            AddParam(dict, "eval", Eval);
+            AddParam(dict, "accessibility_analysis", AccessibilityAnalysis);
 
             AddParam(dict, "transformation", GetTransformation(Transformation));