+++ Shared/Provisioning/Actions/UserResult.cs
@@ -2,6 +2,7 @@ namespace CloudinaryDotNet.Actions
 {
     using System;
     using System.Runtime.Serialization;
+    using Newtonsoft.Json;
 
     /// <summary>
     /// Detailed information about user.
@@ -50,6 +51,7 @@ namespace CloudinaryDotNet.Actions
         /// <summary>
         /// Gets or sets a list of sub-accounts to which the user has access.
         /// </summary>
+        [JsonConverter(typeof(SafeArrayConverter))]
         [DataMember(Name = "sub_account_ids")]
         public string[] SubAccountIds { get; set; }